% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_model.R
\name{add_model}
\alias{add_model}
\title{Set the architecture of models to be trained in deepdive}
\usage{
add_model(
  config = NULL,
  lstm_nodes = NULL,
  dense_nodes = NULL,
  model_name = ""
)
}
\arguments{
\item{config}{\code{character}. The name of the configuration object, created
using create_config()`, that will be edited.}

\item{lstm_nodes}{\code{numeric}. A numerical 'vector' designating the number
of nodes and layers to be used in the long short-term memory architecture.}

\item{dense_nodes}{\code{numeric}. A numerical 'vector' designating the
number of nodes and layers to be used in the dense node architecture.}

\item{model_name}{\code{string}. An identifier to destinguish between models.}
}
\value{
Adds attributes for a new model, specified with lstm_nodes and
dense_nodes.
}
\description{
A function to assign the number of layers and nodes in the long short-term
memory and dense layers of a deep neural network in DeepDive.
}
\details{
This function can be used to train multiple models in the DeepDive
framework. Without running the function, a default of two lstm layers 64, 32
and two fully-connected dense node layers 64, 32 will be used.
}
\examples{
# Import internal dataset
data(carnivora)
# Generate vector describing time bin boundaries
bins <- c(66, 23, 2.6, 0)
# Create configuration object
config <- create_config(name = "carnivora",
                        data_file = "data/carnivora_deepdive_input.csv",
                        bins = bins,
                        n_regions = length(unique(carnivora$Region)))
# Add model architectures to the config file
add_model(config = config, lstm_nodes = c(128, 64), dense_nodes = c(64, 32), name = "1")
}
