% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_bins.R
\name{adjust_epochs}
\alias{adjust_epochs}
\title{Adjust epochs time bins}
\usage{
adjust_epochs(bins, finish, rmv_bins = NULL, quat = F)
}
\arguments{
\item{bins}{A table of time bin start, end and midpoint ages.}

\item{finish}{Ending age of the youngest time bin.}

\item{rmv_bins}{Named bins will be removed from the bins object.}

\item{quat}{When TRUE bins in the Quaternary are merged into one time bin.}
}
\value{
A table of start, end and midpoint ages
}
\description{
'adjust_epochs()' takes a table of time bin start, end and midpoint ages and
makes adjustments depending on settings provided.
}
\examples{
adjust_epochs(bins, finish=192.9, rmv_bins=)
}
