% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_ages.R
\name{ages}
\alias{ages}
\title{Ages function}
\usage{
ages(dat, method)
}
\arguments{
\item{dat}{Occurrence data table}

\item{method}{Age assignment method ("median", "random" or
"random_by_loc")}
}
\value{
Occurrence data table with additional SampledAge column
}
\description{
'ages()' assigns sampled ages to each occurrence within it's age range using
either the median of the minimum and maximum age estimate, a random age drawn
from a uniform distribution, or a random age drawn from uniform distribution
that is kept the same for occurrences which are sampled from the same
locality
}
\examples{
ages(dat, method="median")
}
