% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolution.R
\name{assign_res}
\alias{assign_res}
\title{Assign resolution}
\usage{
assign_res(dat, age_range_threshold = NA, all_lr = F)
}
\arguments{
\item{dat}{Data frame of occurrences with taxon ID, Area, MinAge, MaxAge and
Locality columns.}

\item{age_range_threshold}{Specify an age range cut-off between high and low
resolution fractions (default=NA)}

\item{all_lr}{When set to TRUE all occurrences are assigned LR. Consider
using if there is a very high frequency of low resolution occurrences in a
given data set (default FALSE).}
}
\value{
dat_res, the dat table with an additional column where the
resolution of occurrences are categorised as LR (low res) or HR (high res).
}
\description{
'assign_res()' creates an additional column on the dat object that identifies
an occurrences as having an age range entirely encompassed within a high
resolution time bin (HR) or exceeding the boundaries of high resolution time
bins (LR).
}
\examples{
assign_res(dat)
assign_res(dat, age_range_threshold=NA, all_lr=T)
}
