% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_bins.R
\name{build_bins}
\alias{build_bins}
\title{Build equal length time bins}
\usage{
build_bins(start = 100, end = 0, n = 100)
}
\arguments{
\item{start}{\code{numeric}. Numerical age at which the oldest time bin
should start. Default is \code{100}.}

\item{end}{\code{numeric}. Numerical age at which the youngest time bin
should end. Default is \code{0}.}

\item{n}{\code{numeric}. Number of time bins to be made. Default is \code{100}.}
}
\value{
A \code{dataframe} containing the start, end, and midpoint ages of the
generated time bins.
}
\description{
A function to create a table of time bins. The function is called within the
\code{time_bins()} function only in the case where \code{equal_bins} are specified by
the user. By default, 100 1My bins are generated.
}
\examples{
# Generate 100 time bins between 100My and present
example1 <- build_bins()
# Generate 20 time bins between 10My and 5My
example2 <- build_bins(start = 10, end = 5, n = 20)
}
