% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_bins.R
\name{build_stages}
\alias{build_stages}
\title{Build time bins using geological stages}
\usage{
build_stages(start = 66, end = 0)
}
\arguments{
\item{start}{\code{numeric}. Numerical age at which the oldest time bin
should start. Default is \code{66}.}

\item{end}{\code{numeric}. Numerical age at which the youngest time bin
should end. Default is \code{0}.}
}
\value{
A \code{dataframe} containing the start, end, and midpoint ages of the
generated time bins.
}
\description{
A function to create a table of time bins. The function is called within the
\code{time_bins()} function only in the case where \code{stages} are specified by the
user. By default, Cenozoic stages are returned.
}
\examples{
# Generate time bins for the Cenozoic
example1 <- build_stages()
# Generate time bins for the Mesozoic
example2 <- build_stages(start = 252, end = 66)
}
