% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carnivora}
\alias{carnivora}
\title{Carnivore data}
\format{
A data frame with 11,309 rows and 5 columns:
\describe{
\item{Taxon}{The taxonomic name given to the occurrence.}
\item{Area}{The discrete geographic region in which the taxon was found.}
\item{MinAge}{The youngest age estimate of the fossil occurrence, in
millions of years.}
\item{MaxAge}{The oldest age estimate of the fossil occurrence, in millions
of years.}
\item{Locality}{A unique identifier for the locality at which the fossil
was recorded.}
}
}
\source{
Faurby S., Werdelin L., and Antonelli, A. 2019. Dispersal ability
predicts evolutionary success among mammalian carnivores. bioRxiv,
10.1101, 755207.
\url{https://www.biorxiv.org/content/10.1101/755207v1.abstract}
}
\usage{
carnivora
}
\description{
A subset of occurrence data for the order Carnivora from Faurby et al.
(2019).
}
\keyword{datasets}
