% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_ascent_gas_tbl.R
\name{Create_ascent_gas_tbl}
\alias{Create_ascent_gas_tbl}
\title{Create the ascent gas tibble}
\usage{
Create_ascent_gas_tbl(Gas_list, Max_depth)
}
\arguments{
\item{Gas_list}{A 3-colum tibble with proportions of O2, N2 and He in percent.}

\item{Max_depth}{A double. Maximum depth of the dive}
}
\value{
A tibble containing one row for each segment of the ascent
}
\description{
Create the ascent gas tibble
}
