% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_bottom_gas_tbl.R
\name{Create_bottom_gas_tbl}
\alias{Create_bottom_gas_tbl}
\title{This function creates a tibble containing all relevant informations about the bottom gas
Desc_gas: a tibble as produced by Create_desc_gas_tbl()}
\usage{
Create_bottom_gas_tbl(Desc_gas)
}
\description{
This function creates a tibble containing all relevant informations about the bottom gas
Desc_gas: a tibble as produced by Create_desc_gas_tbl()
}
