% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_desc_gas_tbl.R
\name{Create_desc_gas_tbl}
\alias{Create_desc_gas_tbl}
\title{This function produces a tibble of gases to breathe during the descent
Gas_list: a tibble of 5 columns: O2, N2, He, Min_OD and Max_OD
Returns a tibble with two additional columns (Start_depth and End_depth) and only relevant gases sorted in the correct order for the descent}
\usage{
Create_desc_gas_tbl(Gas_list, Max_depth)
}
\description{
This function produces a tibble of gases to breathe during the descent
Gas_list: a tibble of 5 columns: O2, N2, He, Min_OD and Max_OD
Returns a tibble with two additional columns (Start_depth and End_depth) and only relevant gases sorted in the correct order for the descent
}
