% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepDiveR-package.R
\docType{package}
\name{DeepDiveR-package}
\alias{DeepDiveR}
\alias{DeepDiveR-package}
\title{DeepDiveR: Set and launch DeepDive analyses.}
\description{
Organise input data and scripts for analysis in DeepDive.
}
\author{
\strong{Maintainer}: Rebecca Cooper \email{rebecca.cooper@unifr.ch} (\href{https://orcid.org/0000-0002-4742-1422}{ORCID})

Authors:
\itemize{
  \item Bethany Allen \email{bethany.allen@bsse.ethz.ch} (\href{https://orcid.org/0000-0003-0282-6407}{ORCID})
  \item Daniele Silvestro \email{daniele.silvestro@unifr.ch} (\href{https://orcid.org/0000-0003-0100-0961}{ORCID})
}

}
\keyword{internal}
