% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_bins.R
\name{div_bins}
\alias{div_bins}
\title{Make time bins which line up with geological stages or epochs}
\usage{
div_bins(
  bin_type,
  begin,
  finish,
  t_res = NULL,
  n_bins = NA,
  remove_bins = NULL,
  use_q = FALSE,
  merge_holo = FALSE,
  merge_holo_ple = FALSE
)
}
\arguments{
\item{bin_type}{Determines how bins will be made (equal bins, epochs, or stages)}

\item{begin}{Starting age of the oldest time bin}

\item{finish}{Ending age of the youngest time bin}

\item{t_res}{Adjust the average length of time bins (t_res = 1 gives bins averaging 1 Ma in length)}

\item{n_bins}{Number of bins (used only when equal_bins are specified)}

\item{remove_bins}{Specify named time bins to be removed (quote directly or
use a list object)}

\item{use_q}{Merge time bins in the Quaternary together}

\item{merge_holo}{Merge time bins in the Holocene together}

\item{merge_holo_ple}{Merge time bins in the Upper Pleistocene and Holocene}

\item{res}{Return "high" or "low" resolution time bins}
}
\value{
A table of time bins with start, end and midpoint ages. Where epochs
or stages specified names are stored as well.
}
\description{
'div_bins()' makes a table of time bin start, end and midpoint ages at
either the scale of epochs or stages, which are then broken down into smaller
time bins of specified average length with bin boundaries that align to those
of the geological stages or epochs.
}
\examples{
div_bins(bin_type="stages", begin=264.28 , finish=192.9, t_res=NULL)
div_bins(bin_type="stages", begin=264.28, finish=192.9, t_res=1)
div_bins(bin_type="stages", begin=264.28 , finish=192.9, t_res=10)
}
