% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_config.R
\name{edit_config}
\alias{edit_config}
\title{Edit values in the configuration file}
\usage{
edit_config(config = NULL, module = NULL, parameter = NULL, value = NULL)
}
\arguments{
\item{config}{\code{character}. The name of the configuration object, created
using \code{create_config()}, that will be edited.}

\item{module}{\code{character}. The name of the block of the configuration
object to which the parameter belongs.}

\item{parameter}{\code{character}. The name of the parameter to edit in
the configuration file.}

\item{value}{\code{}. The value you want to assign to the parameter. The type
of value is dependent upon the parameter.}
}
\value{
A configuration object with the value of the desired parameter
changed.
}
\description{
A function to modify parameters within a configuration object for DeepDive,
created using \code{create_config()}. Default values are listed below, while
the values currently held by each parameter can be checked using the
notation \verb{config$data$[module]$[parameter]}. Extensive checks ensure that
it is possible for the specific parameter to hold the desired value.
}
\details{
Add full list of default parameters.
}
\examples{
edit_config(config = config, module = "simulations", parameter = "extant_sp",
    value = c(10, 10000))
}
