% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolution.R
\name{fraction_lr_hr_dat}
\alias{fraction_lr_hr_dat}
\title{What fraction of the data is at high vs low resolution}
\usage{
fraction_lr_hr_dat(dat, res = "low", r_dat = NA, age_range_threshold = NA)
}
\arguments{
\item{dat}{Occurrence data table}

\item{res}{Specify "high" or "low" fraction to be returned}

\item{r_dat}{Occurrence data table at either high or low resolution}

\item{age_range_threshold}{Specify an age range cut-off between high and low
resolution fractions (default=NA)}
}
\value{
Returns fraction of the occurrences at either high or low resolution
depending on parameters used.
}
\description{
'fraction_lr_hr_dat()' finds the fraction of occurrences which have an age
range that is entirely within either the specified low resolution or high
resolution time bins.
}
\examples{
fraction_lr_hr_dat(dat, r_dat = lr_dat, res="low")
}
