% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localities.R
\name{generate_locality_dataset}
\alias{generate_locality_dataset}
\title{Summarise locality distribution}
\usage{
generate_locality_dataset(dat = NULL, bins = NULL)
}
\arguments{
\item{dat}{\code{dataframe}. A \code{dataframe} containing the fossil occurrences,
including \code{Taxon}, \code{Area}, \code{MinAge}, \code{MaxAge} and \code{Locality} columns.}

\item{bins}{\code{dataframe}. A \code{dataframe} designating the boundaries of
the time bins used in the analysis. These should reflect the empirical data
and be identical to the bins used in any corresponding DeepDive simulations
and training.}
}
\value{
A \code{dataframe} of shape \verb{region x time bin} describing where
occurrences are recorded.
}
\description{
A function to create a \code{dataframe} of the summed number of localities per
area and time bin, provided as a table of regions x bins.
}
\examples{
# Import internal dataset
dat <- tetrapods
# Set Carboniferous to Triassic time bins
time_bins <- build_stages(start = 359, end = 200)
# Summarise distribution of occurrences
example1 <- generate_locality_dataset(dat = dat, bins = time_bins)
}
