% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occs_per_region_time.R
\name{generate_occurrence_dataset}
\alias{generate_occurrence_dataset}
\title{Summarise occurrences per continent and time bin}
\usage{
generate_occurrence_dataset(dat = NULL, bins = NULL)
}
\arguments{
\item{dat}{\code{dataframe}. A \code{dataframe} containing the fossil occurrences,
including \code{Taxon}, \code{Area}, \code{MinAge}, \code{MaxAge} and \code{Locality} columns. A
\code{SampledAge} column must also be generated, such as using \code{ages()}.}

\item{bins}{\code{dataframe}. A \code{dataframe} designating the boundaries of
the time bins used in the analysis. These should reflect the empirical data
and be identical to the bins used in any corresponding DeepDive simulations
and training.}
}
\value{
A \code{dataframe} of shape \verb{region x time bin} describing where
occurrences are recorded.
}
\description{
A function to summarise the number of occurrences per region, provided as a
table of regions x bins.
}
\examples{
#' # Import internal dataset
dat <- tetrapods
#' # Assign ages for tetrapods
dat <- ages(dat = tetrapods)
# Set Carboniferous to Triassic time bins
time_bins <- build_stages(start = 359, end = 200)
# Summarise distribution of occurrences
example1 <- generate_occurrence_dataset(dat = dat, bins = time_bins)
}
