% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geo_bins}
\alias{geo_bins}
\title{Geological stage data}
\format{
A data frame with 102 rows and 9 columns:
\describe{
\item{sys}{The geological period containing the stage.}
\item{series}{The geological series containing the stage.}
\item{stage}{The name of the geological stage.}
\item{start}{The age of the start of the geological stage, in millions of
years.}
\item{end}{The age of the end of the geological stage, in millions of
years.}
\item{midpoint}{The age of the midpoint of the geological stage, in
millions of years.}
\item{dur}{The duration of the geological stage, in millions of years.}
\item{seriescol}{The hexadecimal colour usually given to the geological
series.}
\item{stagecol}{The hexadecimal colour usually given to the geological
stage.}
}
}
\source{
#To be completed#
}
\usage{
geo_bins
}
\description{
A description of the ages given to all Phanerozoic geological stages.
}
\keyword{datasets}
