% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolution.R
\name{get_lr_hr_dat}
\alias{get_lr_hr_dat}
\title{Split low and high resolution data     ########## THIS IS NO LONGER IN USE BUT CONSIDER COMBINING WITH ASSIGN_RES SO THERE ARE OPTIONS TO STILL USE AN AGE_RANGE_THRESHOLD}
\usage{
get_lr_hr_dat(data = dat, res = "high", age_range_threshold = NA)
}
\arguments{
\item{data}{Occurrence data table}

\item{res}{Specify whether to return "low" or "high" resolution data}

\item{age_range_threshold}{Specify an arbitrary cut off point at which to
split high and low resolution data (default = NA) - not based on bin
boundaries in Ma.}
}
\value{
Data frame of occurrences at either high or low resolution, with
the shape of the data frame held constant.
}
\description{
'get_lr_hr_dat()' splits occurrence data by resolution, where low resolution
data comprises occurrences with age ranges that do not cross low resolution
bin boundaries and high resolution data that exceed the low resolution bin
boundaries.
}
\examples{
get_lr_hr_dat(data = dat_res, res = "low", age_range_threshold = NA)
get_lr_hr_dat(data = dat_res, res = NA, age_range_threshold = 2)
}
