% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localities_through_time.R
\name{localities_through_time}
\alias{localities_through_time}
\title{Summarise locality occupation through time}
\usage{
localities_through_time(dat = NULL, bins = NULL)
}
\arguments{
\item{dat}{\code{dataframe}. A \code{dataframe} containing the fossil occurrences,
including \code{Taxon}, \code{Region}, \code{Locality} and \code{SampledAge} columns. The
supplied \code{dataframe} should not contain any \code{NA} values.}

\item{bins}{\code{numeric}. A numerical \code{vector} designating the boundaries
of the time bins used in the analysis.}
}
\value{
A \code{dataframe} of shape 'region x time bin' describing where
occurrences are recorded.
}
\description{
A function to create a \code{dataframe} of the number of localities per region and
time bin.
}
\examples{
# Import internal dataset
data(carnivora)
# Generate vector describing time bin boundaries
bins <- c(66, 23, 2.6, 0)
# Generate sampled ages using the "median" method
carnivora <- ages(dat = carnivora)
# Compute locality counts
localities_through_time(dat = carnivora, bins = bins)
}
