% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occs_area_time.R
\name{occs_area_time}
\alias{occs_area_time}
\title{Summarise occurrences per area per time bin}
\usage{
occs_area_time(dat = NULL, bins = NULL)
}
\arguments{
\item{dat}{\code{dataframe}. The \code{dataframe} describing the occurrence data,
including \code{Taxon}, \code{Area}, \code{Locality} and \code{SampledAge} columns. The
supplied \code{dataframe} should not contain any \code{NA} values.}

\item{bins}{\code{numeric}. A numerical \code{vector} designating the boundaries
of the time bins used in the analysis.}
}
\value{
A \code{dataframe} of shape 'region x time bin' describing the number of
occurrences in each category.
}
\description{
A function to produce a \code{dataframe} of the number of occurrences found in
each geographic area in each time bin.
}
\examples{
# Import internal dataset
data(carnivora)
# Generate vector describing time bin boundaries
bins <- c(66, 23, 2.6, 0)
# Generate sampled ages using the "median" method
carnivora <- ages(dat = carnivora)
# Compute occurrences per area and time
occs_area_time(dat = carnivora, bins = bins)
}
