% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Operating_depth.R
\name{Operating_depth}
\alias{Operating_depth}
\title{Operating_depth computes the maximum (default, limit = "bottom") and minimum (when limit = "top") operating depth for one or more gases.
O2: tibble, vector or scalar. Proportion of O2 in the mix in percent.
PO2: double. PO2 for which the limit is to be calculated (typically, 0.18 or 1.61)
limit: character. One of "bottom" or "top". Should be set to "bottom" (default) for Maximum OD, and to "top" for Minimum OD.}
\usage{
Operating_depth(O2, PO2, limit = c("bottom", "top"))
}
\description{
Operating_depth computes the maximum (default, limit = "bottom") and minimum (when limit = "top") operating depth for one or more gases.
O2: tibble, vector or scalar. Proportion of O2 in the mix in percent.
PO2: double. PO2 for which the limit is to be calculated (typically, 0.18 or 1.61)
limit: character. One of "bottom" or "top". Should be set to "bottom" (default) for Maximum OD, and to "top" for Minimum OD.
}
