% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parameters}
\alias{parameters}
\title{DeepDive configuration file parameters}
\format{
A data frame with 80 rows and 6 columns:
\describe{
\item{module}{The module in the configuration file within which the
parameter can be found.}
\item{parameter}{The name of the parameter.}
\item{description}{A brief description of the parameter.}
\item{default}{The default value given to the parameter when the
configuration file is created.}
\item{autotune}{A logical value describing whether the parameter can be
autotuned (TRUE) or not (FALSE).}
\item{limits}{The type of value which can be attributed to the parameter.}
}
}
\usage{
parameters
}
\description{
A table describing the parameters included in the configuration file.
}
\keyword{datasets}
