% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_comp.R
\docType{data}
\name{period}
\alias{period}
\title{Computes the partial ppressure of a gaz in each of the 16 tissues defined by Bühlmann (table ZH-L16 C)
half.t est la période du tissus. Peut être une unique valeur ou un vecteur
prof.init est la profondeur initiale en msw
prof.fin est la profonder finale en msw
La variation de profondeur doit se faire é vitesse constante (en msw / min)
La variable gaz correspond é la proportion du gaz considéré dans le mélange (N2 ou He)
La variable molec permet d'indiquer si l'on souhaite le profil de l'azote ou de l'hélium
A la remontée, indiquer descente = F
La variable dur.qr correspond é la durete du modéle prenant en compte la pression partielle de vapeur d'eau dans les alvéoles. Elle peut prendre les valeurs 3, 2 ou 1' qui correspondent respectivement é 0.8, la valeur plus stricte, (Schreiner), 0.9 (US Navy) ou 1 (la plus permissive, Béhlmann).
é profondeur constante (prof.init == prof.fin), la durée t doit étre renseignée car elle ne peut étre déduite de la vitesse et de la différence de profondeur
Pprec est la pression partielle en gaz inerte dans les tissus é la fin de l'étape précédente. Si on part de la surface, elle vaut 0 pour l'hélium et la valeur de saturation pour l'azote, sinon, elle vaut les valeurs de pressions calculées par la fonction é l'issue du tréon précédent}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 32 rows and 2 columns.}
\usage{
period
}
\description{
Computes the partial ppressure of a gaz in each of the 16 tissues defined by Bühlmann (table ZH-L16 C)
half.t est la période du tissus. Peut être une unique valeur ou un vecteur
prof.init est la profondeur initiale en msw
prof.fin est la profonder finale en msw
La variation de profondeur doit se faire é vitesse constante (en msw / min)
La variable gaz correspond é la proportion du gaz considéré dans le mélange (N2 ou He)
La variable molec permet d'indiquer si l'on souhaite le profil de l'azote ou de l'hélium
A la remontée, indiquer descente = F
La variable dur.qr correspond é la durete du modéle prenant en compte la pression partielle de vapeur d'eau dans les alvéoles. Elle peut prendre les valeurs 3, 2 ou 1' qui correspondent respectivement é 0.8, la valeur plus stricte, (Schreiner), 0.9 (US Navy) ou 1 (la plus permissive, Béhlmann).
é profondeur constante (prof.init == prof.fin), la durée t doit étre renseignée car elle ne peut étre déduite de la vitesse et de la différence de profondeur
Pprec est la pression partielle en gaz inerte dans les tissus é la fin de l'étape précédente. Si on part de la surface, elle vaut 0 pour l'hélium et la valeur de saturation pour l'azote, sinon, elle vaut les valeurs de pressions calculées par la fonction é l'issue du tréon précédent
}
\keyword{datasets}
