% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_dd_input}
\alias{plot_dd_input}
\title{Plot deepdive_input}
\usage{
plot_dd_input(dd_input)
}
\arguments{
\item{dd_input}{Deepdive_input object made by the function 'prep_dd_input()'
which comprises the duration and midpoints of time bins, the number of
localities through time per area and the number of occurrences through time
for each species per area.}
}
\value{
A file containing ___________ plots.
}
\description{
'dd_inputs_plot()' makes and saves a variety of plots to start exploring your
input data which will save in the file _____.
}
