% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plt_fads_lads}
\alias{plt_fads_lads}
\title{Plot FADS and LADS}
\usage{
plt_fads_lads()
}
\description{
'plt_fads_lads()' plots the first appearance datum and last appearance datum
of the occurrences through time.
}
