% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepdive_input_files.R
\name{prep_dd_input}
\alias{prep_dd_input}
\title{Prepare DeepDive Input}
\usage{
prep_dd_input(
  dat = dat,
  bins,
  r = 10,
  age_m = "random_by_loc",
  output_file = NULL
)
}
\arguments{
\item{dat}{Data frame of occurrences with taxon ID, Area, MinAge, MaxAge and
Locality columns.}

\item{bins}{A vector of time bins}

\item{r}{The number of age assignment replicates}

\item{age_m}{The age assignment method}

\item{output_file}{A string containing an output file name.}
}
\value{
A .csv file containing: start age, midpoint and duration of time
bins, counts of the number of localities per region for each bin, counts of
occurrences of each taxon per region per time bin, the data type (bins, locs,
occs) and replicate number (where age replicates are used, column R).
}
\description{
'prep_dd_input()' makes a single input file for use in DeepDive in one line.
}
\examples{
prep_dd_input(dat=your_data, bins=bins, r=100, age_m = "random_by_loc")
}
