% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions_matrix.R
\name{regions_matrix}
\alias{regions_matrix}
\title{Make discrete geographic regions appear or disappear}
\usage{
regions_matrix(config = NULL, region_ages = NULL, presence = TRUE)
}
\arguments{
\item{config}{\code{character}. The name of the configuration object, created
using \code{create_config()}, that will be edited.}

\item{region_ages}{\code{dataframe}. If left NULL (default), all geographic
regions are configured to be present throughout the simulations. Otherwise,
a \code{dataframe} describing for each geographic \code{Region} the \code{MinAge}
(younger limit) and \code{MaxAge} (older limit) between which a geographic
connection is either made or severed. The number of rows in the \code{dataframe}
must match \code{n_regions} in the configuration file. The supplied \code{dataframe}
should not contain any \code{NA} values.}

\item{presence}{\code{logical}. When TRUE (the default), regions become
available to occupy based on the time frames specified. When FALSE, regions
will be removed from the simulations based on the time frames specified.}
}
\value{
Adds an attribute for each region (named "Region1", "Region2"...
etc.) with minimum and maximum certainty limits of migration becoming
possible.
}
\description{
A function to add area constraints to new module in the configuration file
by providing a maximum and minimum age uncertainty for geographic regions
being connected to or disconnected from the study system.
}
\details{
Note that this function does not need to be used if all geographic
regions are considered to be accessible throughout the simulations; this is
the default.
}
\examples{
# Import internal dataset
data(carnivora)
# Generate vector describing time bin boundaries
bins <- c(66, 23, 2.6, 0)
# Create configuration object
config <- create_config(name = "carnivora",
                        data_file = "data/carnivora_deepdive_input.csv",
                        bins = bins,
                        n_regions = length(unique(carnivora$Region)))
# Generate dataframe describing region presence
region_ages <- rbind(c("Africa", 66, 66),
                   c("Asia", 66, 66),
                   c("Europe", 66, 66),
                   c("North America", 66, 66),
                   c("South America", 11.608, 7.3))
region_ages <- as.data.frame(region_ages)
# Label columns
colnames(region_ages) <- c("Region", "MaxAge", "MinAge")
# Connect region data to configuration file
region_matrix(config = config, region_ages = region_ages)
}
