% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_config.R
\name{set_value}
\alias{set_value}
\title{Edit values in config}
\usage{
set_value(attribute_name, value, module, config)
}
\arguments{
\item{attribute_name}{A string providing direction to the attribute you want to edit e.g. "total_sp"}

\item{value}{An integer or vector value you want to assign to the attribute.}

\item{module}{A string indicating the block of the config the attribute can be found within.}

\item{config}{A config which can be generated in create_config(), that you will edit here.}
}
\value{
Edits settings in the config .ini file.
}
\description{
'set_value()' accesses attributes in the config file and adjusts the assigned
value(s) to that of a provided integer or vector.
}
\examples{
set_value(attribute_name = "extant_sp", value=c(10, 10000), module="simulations", config)
}
