% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_occurrences.R
\name{taxa_time_per_area}
\alias{taxa_time_per_area}
\title{Calculate the number of taxa per time bin per area}
\usage{
taxa_time_per_area(dat = NULL, bins = NULL)
}
\arguments{
\item{dat}{\code{dataframe}. The \code{dataframe} describing the occurrence data,
including \code{Taxon}, \code{Area}, \code{MinAge}, \code{MaxAge} and \code{Locality} columns. A
\code{SampledAge} column must also be generated, such as using \code{ages()}.}

\item{bins}{\code{dataframe}. A \code{dataframe} designating the boundaries of
the time bins used in the analysis. These should reflect the empirical data
and be identical to the bins used in any corresponding DeepDive simulations
and training.}
}
\value{
A \code{dataframe} of shape \verb{taxon x region x time bin} describing where
occurrences are recorded.
}
\description{
A function to produce a table of shape \verb{taxon x region x time bin} where
occurrences are recorded.
}
\examples{
# Import internal dataset
dat <- tetrapods
# Assign ages for tetrapods
dat <- ages(dat = tetrapods)
# Set Carboniferous to Triassic time bins
time_bins <- build_stages(start = 359, end = 200)
# Calculate number of taxa per time bin and region
example1 <- taxa_time_per_area(dat = tetrapods, bins = time_bins)
}
