% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_bins.R
\name{time_bins_all_opts}
\alias{time_bins_all_opts}
\title{Make time bins all options}
\usage{
time_bins_all_opts(
  bin_type,
  begin,
  finish,
  n_bins = 100,
  custom_bins = NULL,
  lr_hr_bins = "",
  remove_bins = NULL,
  use_q = FALSE,
  merge_holo = FALSE,
  merge_holo_ple = FALSE
)
}
\arguments{
\item{bin_type}{Determines how bins will be made (equal bins, epochs, or
stages) or whether "custom_bins" will be used.}

\item{begin}{Starting age of the oldest time bin}

\item{finish}{Ending age of the youngest time bin}

\item{n_bins}{Number of bins (used only when equal_bins are specified)}

\item{lr_hr_bins}{Make time bins at a specified low or high resolution}

\item{remove_bins}{Specify named time bins to be removed (quote directly or
use a list object)}

\item{use_q}{Merge time bins in the Quaternary together}

\item{merge_holo}{Merge time bins in the Holocene together}

\item{merge_holo_ple}{Merge time bins in the Upper Pleistocene and Holocene}
}
\value{
A table of time bins with start, end and midpoint ages. Where epochs
or stages specified names are stored as well.
}
\description{
'time_bins_all_opts()' makes a table of time bin start, end and midpoint ages
}
\examples{
time_bins_all_opts(bin_type="equal_bins", begin=100, finish=0, n_bins=10)
time_bins_all_opts(bin_type="stages", begin=251.902, finish=66)
time_bins_all_opts(bin_type="epochs", begin=251.902, finish=66)
time_bins_all_opts(bin_type="stages", begin=66, finish=0, remove_bins="Danian")
time_bins_all_opts(bin_type="stages", begin=66, finish=0, use_q = T)
time_bins_all_opts(bin_type="stages", begin=66, finish=0, merge_holo = T)
time_bins_all_opts(bin_type="stages", begin=66, finish=0, merge_holo_ple = T)
time_bins_all_opts(bin_type="epochs", begin=66, finish=0, use_q = T)
time_bins_all_opts(bin_type="epochs", begin=66, finish=0, remove_bins = "Holocene")
time_bins_all_opts(bin_type="custom_bins", custom_bins = -c(10:0))
time_bins_all_opts(bin_type="stages", begin=66, finish=0, lr_hr_bins="low")
time_bins_all_opts(bin_type="stages", begin=66, finish=0, lr_hr_bins="high")
}
