% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_bins.R
\name{adjust_stages}
\alias{adjust_stages}
\title{Adjust stages time bins}
\usage{
adjust_stages(bins, finish, rmv_bins = NULL, quat = F, holo = F, holo_ple = F)
}
\arguments{
\item{bins}{A table of time bin start, end and midpoint ages.}

\item{finish}{Ending age of the youngest time bin.}

\item{rmv_bins}{Named bins will be removed from the bins object.}

\item{quat}{When TRUE bins in the Quaternary are merged into one time bin.}

\item{holo}{When TRUE bins in the Holocene will be merged into one time bin.}

\item{holo_ple}{When TRUE bins in the Holocene will be merged into one time
bin along with the Upper Pleistocene.}
}
\value{
A table of start, end and midpoint ages
}
\description{
'adjust_stages()' takes a table of time bin start, end and midpoint ages and
makes adjustments depending on settings provided.
}
\examples{
adjust_stages(bins, finish=192.9 , rmv_bins="Sinemurian")
adjust_stages(bins, finish=0, quat=T)
}
