% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areas_matrix.R
\name{areas_matrix}
\alias{areas_matrix}
\title{Make discrete geographic regions appear or disappear}
\usage{
areas_matrix(config = NULL, area_ages = NULL, presence = TRUE)
}
\arguments{
\item{config}{\code{character}. The name of the configuration object, created
using \code{create_config()}, that will be edited.}

\item{area_ages}{\code{dataframe}. If left NULL (default), all geographic
regions are configured to be present throughout the simulations. Otherwise,
a \code{dataframe} describing the \code{MinAge} and \code{MaxAge} for each geographic
\code{Area}, with older ages in column one. The number of rows in the
\code{dataframe} must match \code{n_areas} in the configuration file. The
supplied \code{dataframe} should not contain any \code{NA} values.}

\item{presence}{\code{logical}. When TRUE (the default), regions become
available to occupy within the time frames specified. When FALSE, regions
will be removed from the simulations between the ages specified.}
}
\value{
Adds an attribute for each area (named "area1", "area2"... etc) with
minimum and maximum age of migration becoming possible.
}
\description{
A function to modify the simulations module by providing a maximum and
minimum age range for geographic regions being connected to the study
system.
}
\examples{
# Generate dataframe describing area presence
area_ages <- rbind(c("Africa", 66, 66),
                   c("Asia", 66, 66),
                   c("Europe", 66, 66),
                   c("North America", 66, 66),
                   c("South America", 11.608, 7.3))
 area_ages <- as.data.frame(area_ages)

areas_matrix(config = config, area_ages = area_ages)
}
