% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_bins.R
\name{build_bins}
\alias{build_bins}
\title{Build equal length time bins}
\usage{
build_bins(start, end, n = NULL)
}
\arguments{
\item{start}{Age at which the oldest time bin should start.}

\item{end}{Age at which the youngest time bin should end.}

\item{n}{Number of bins to be made.}
}
\value{
Tibble containing the start, end and midpoint ages of time bins.
}
\description{
'build_bins()' makes a table of time bins and is called within the time_bins
function only in the case where equal_bins are specified by the user.
}
\examples{
build_bins(start=100, end=0, n=100)
build_bins(start=10, end=5, n=20)
}
