% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_bins.R
\name{build_epochs}
\alias{build_epochs}
\title{Build time bins using geological epochs}
\usage{
build_epochs(begin, finish)
}
\arguments{
\item{start}{Age at which the oldest time bin should start.}

\item{end}{Age at which the youngest time bin should end.}
}
\value{
Table containing the start, end and midpoint ages of time bins.
}
\description{
'build_epochs()' makes a table of time bins and is called within the
time_bins function only in the case where epochs are specified by the user.
}
\examples{
build_epochs(start=251.902, end=66)
}
