% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_bins.R
\name{build_stages}
\alias{build_stages}
\title{Build time bins using geological stages}
\usage{
build_stages(begin, finish)
}
\arguments{
\item{begin}{Age at which the oldest time bin should start.}

\item{finish}{Age at which the youngest time bin should end.}
}
\value{
Table containing the start, end and midpoint ages of time bins.
}
\description{
'build_stages()' makes a table of time bins and is called within the
time_bins function only in the case where stages are specified by the user.
}
\examples{
build_stages(start=66, end=0)
}
