% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carnivora}
\alias{carnivora}
\title{Carnivore data}
\format{
A data frame with 11,309 rows and 5 columns:
\describe{
\item{Taxon}{The taxonomic name given to the occurrence.}
\item{Area}{The discrete geographic region in which the taxon was found.}
\item{MinAge}{The youngest age estimate of the fossil occurrence, in
millions of years.}
\item{MaxAge}{The oldest age estimate of the fossil occurrence, in millions
of years.}
\item{Locality}{A unique identifier for the locality at which the fossil
was recorded.}
}
}
\source{
Faurby S., Silvestro D., Werdelin L., and Antonelli, A. 2024. Reliable
biogeography requires fossils: insights from a new species-level phylogeny of
extinct and living carnivores. Proceedings of the Royal Society B,
291(2028), 20240473.
\url{https://royalsocietypublishing.org/doi/full/10.1098/rspb.2024.0473t}
}
\usage{
carnivora
}
\description{
A subset of occurrence data for the order Carnivora from Faurby et al.
(2024).
}
\keyword{datasets}
