% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_config.R
\name{create_config}
\alias{create_config}
\title{Create DeepDive config object and file}
\usage{
create_config(
  simulate = T,
  model_training = T,
  empirical_predictions = F,
  outputs = F,
  autotune = FALSE,
  path_wd = NULL,
  time_bins = NULL,
  sim_name = NULL,
  n_areas = NULL,
  simulations_file = NULL,
  add_test = F,
  models_file = NULL,
  feature_file = NULL,
  label_file = NULL,
  empirical_input_file = dd_dataset,
  include_present_diversity = FALSE,
  present_diversity = NA,
  calibrate_diversity = FALSE,
  taxonomic_level = taxon_level,
  output_file = NULL
)
}
\arguments{
\item{simulate}{A true or false statement that switches on and off the generation of simulated datasets used in model training and testing.}

\item{model_training}{A true or false statement that switches on and off the training of new RNN models.}

\item{empirical_predictions}{A true or false statement that switches on and off the set up for empirical analyses.}

\item{path_wd}{The working directory where files for analyses can be found.}

\item{time_bins}{A vector of bin boundary ages.}

\item{sim_name}{A string that identifies a set of simulations.}

\item{n_areas}{An integer number of discrete sampling regions e.g. continents, basins.}

\item{simulations_file}{File where simulations will be saved.}

\item{add_test}{A true or false statement, when true a test set with the same settings as the training set are generated.}

\item{models_file}{File where trained models can be saved and retrieved.}

\item{feature_file}{File name for training features.}

\item{label_file}{File name for training labels.}

\item{empirical_input_file}{File name where empirical data in the input format for use in deepdive will be saved.}

\item{include_present_diversity}{A true or false statement, if true predictions will be conditioned on modern diversity.}

\item{taxonomic_level}{String of the name of the taxonomic data column, e.g. "Species", "Genus".}

\item{output_file}{File where outputs will be saved.}

\item{present_diverstiy}{A number of extant taxa, which predictions will be conditioned on.}
}
\value{
Creates config with settings to launch DeepDive.
}
\description{
'create_config()' generates a config object with settings needed to launch
DeepDive scripts in python. Uses the ConfigParser formatting.
}
\examples{
config <- create_config(simulate = T,  model_training = F, 
empirical_predictions = F, wd = paste0(getwd()), time_bins = bins, 
sim_name="test", n_areas = length(unique(dat$Area)), 
simulations_file = paste0(path_dat, "simulations"), add_test = F)
}
