% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_config.R
\name{create_config}
\alias{create_config}
\title{Create DeepDive configuration file}
\usage{
create_config(
  name = NULL,
  data_file = NULL,
  bins = NULL,
  n_regions = 1,
  simulate = TRUE,
  model_training = TRUE,
  empirical_predictions = TRUE,
  autotune = TRUE,
  add_test = TRUE,
  present_diversity = NA
)
}
\arguments{
\item{name}{\code{character}. Name of the analysis, used as a prefix for all
file names linked to an analysis. If multiple analyses are run with the
same name, files are overwritten.}

\item{data_file}{\code{character}. Name of the file containing empirical
input data generated by \code{prep_dd_input()}.}

\item{bins}{\code{numeric}. A \code{vector} of bin boundary ages.}

\item{n_regions}{\code{integer}. The number of discrete sampling regions, e.g.
continents, basins. Defaults to 1.}

\item{simulate}{\code{logical}. A TRUE/FALSE statement that switches on and
off generation of simulated datasets used in model training and testing.
Defaults to TRUE.}

\item{model_training}{\code{logical}. A TRUE/FALSE statement that switches on
and off the training of new RNN models. Defaults to TRUE.}

\item{empirical_predictions}{\code{logical}. A TRUE/FALSE statement that
switches on and off empirical analyses. Defaults to TRUE.}

\item{autotune}{\code{logical}. When TRUE (default), a new configuration file
will be saved with adjusted parameters used for analyses that reflect the
empirical data.}

\item{add_test}{\code{logical}. A TRUE/FALSE statement, when true (the
default) a test set with the same settings as the training set are
generated.}

\item{present_diversity}{\code{numerical}. A numeric value describing the
known diversity at the end of the study interval. In the case a clade is
known to be extinct, this should be set to zero. If the end diversity is
unknown, the default NA can be used.}
}
\value{
Creates configuration object with settings to launch DeepDive. Once
finalised, configuration files can be saved using
\code{config$write("file_name.ini")}.
}
\description{
A function to generate a configuration object containing settings needed to
launch DeepDive analyses in Python. Uses formatting compatible with the
configparser library in Python via the R package ConfigParser (Hoefling, 2017).
The first four arguments describe the data inputs and outputs, while
the last five turn on or off major elements of the analytical pipeline.
The general module is always mandatory.
}
\examples{
# Import internal dataset
data(carnivora)
# Generate vector describing time bin boundaries
bins <- c(66, 23, 2.6, 0)
# Create configuration object
config <- create_config(name = "carnivora",
                        data_file = "data/carnivora_deepdive_input.csv",
                        bins = bins,
                        n_regions = length(unique(carnivora$Region)))
}
