% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_config.R
\name{create_config}
\alias{create_config}
\title{Create DeepDive configuration file}
\usage{
create_config(
  simulate = TRUE,
  model_training = TRUE,
  empirical_predictions = TRUE,
  autotune = TRUE,
  present_diversity = NA,
  path_wd = NULL,
  bins = NULL,
  sim_name = "simulations",
  n_areas = NULL,
  simulations_file = "simulations",
  add_test = TRUE,
  models_file = "trained_models",
  feature_file = NULL,
  label_file = NULL,
  data_file = NULL,
  output_file = NULL
)
}
\arguments{
\item{simulate}{\code{logical}. A TRUE/FALSE statement that switches on and
off generation of simulated datasets used in model training and testing.
Defaults to TRUE.}

\item{model_training}{\code{logical}. A TRUE/FALSE statement that switches on
and off the training of new RNN models. Defaults to TRUE.}

\item{empirical_predictions}{\code{logical}. A TRUE/FALSE statement that
switches on and off empirical analyses. Defaults to TRUE.}

\item{autotune}{\code{logical}. When TRUE (default), a new configuration file
will be saved with adjusted parameters used for analyses that reflect the
empirical data.}

\item{present_diversity}{\code{integer}. The number of extant taxa which
predictions will be conditioned on. When NA (the default), no
conditioning occurs.}

\item{path_wd}{\code{character}. The working directory where files for
analyses can be found. This is the directory in which Python will search
for the input files.}

\item{bins}{\code{numeric}. A vector of bin boundary ages.}

\item{sim_name}{\code{character}. A name string to identify a set of
simulations.}

\item{n_areas}{\code{integer}. The number of discrete sampling regions, e.g.
continents, basins.}

\item{simulations_file}{\code{character}. The name of the file where
simulations will be saved.}

\item{add_test}{\code{logical}. A TRUE/FALSE statement, when true (the
default) a test set with the same settings as the training set are
generated.}

\item{models_file}{\code{character}. The file where trained models will be
saved and retrieved.}

\item{feature_file}{\code{character}. The file name for training features.}

\item{label_file}{\code{character}. The file name for training labels.}

\item{data_file}{\code{character}. The file name containing the empirical
input data for DeepDive, as generated by \code{prep_dd_input()}.}

\item{output_file}{\code{character}. The file where outputs will be saved.}
}
\value{
Creates configuration file with settings to launch DeepDive.
}
\description{
A function to generate a configuration object containing settings needed to
launch DeepDive analyses in Python. Uses ConfigParser (Hoefling, 2017)
formatting.
}
\examples{
config <- create_config(path_wd = paste0(getwd()),
                        bins = bins <- c(66, 23, 2.6, 0),
                        sim_name = "test",
                        n_areas = length(unique(your_data$Area)),
                        simulations_file = paste0(path_dat, "simulations"),
                        add_test = T,
                        present_diversity = 200)
}
