% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_bins.R
\name{customise_bins}
\alias{customise_bins}
\title{Custom time bins}
\usage{
customise_bins(custom_bins)
}
\arguments{
\item{custom_bins}{A negative vector of time bin boundary ages}
}
\value{
A table of start, end and midpoint ages
}
\description{
'customise_bins()' takes a vector of time bin boundary ages (assumed to be -)
and makes a table of start, end and midpoint ages.
}
\examples{
customise_bins(custom_bins = -c(264.26, 224.54, 192.9))
}
