% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_config.R
\name{edit_config}
\alias{edit_config}
\title{Edit values in the configuration file}
\usage{
edit_config(config = NULL, module = NULL, parameter = NULL, value = NULL)
}
\arguments{
\item{config}{\code{character}. The name of the configuration object, created
using \code{create_config()}, that will be edited.}

\item{module}{\code{character}. The name of the block of the configuration
object to which the parameter belongs.}

\item{parameter}{\code{character}. The name of the parameter to edit in
the configuration file.}

\item{value}{\code{}. The value you want to assign to the parameter. The type
of value is dependent upon the parameter.}
}
\value{
A configuration object with the value of the desired parameter
changed.
}
\description{
A function to modify parameters within a configuration object for DeepDive,
created using \code{create_config()}. Default values are listed below, while
the values currently held by each parameter can be checked using the
notation \verb{config$data$[module]$[parameter]}. Extensive checks ensure that
it is possible for the specific parameter to hold the desired value.
}
\details{
A full list of parameters is outlined below.

Parameters relating to geographic area inclusion in the simulations are
better altered using \code{areas_matrix()}.
}
\examples{
# Import internal dataset
data(carnivora)
# Generate vector describing time bin boundaries
bins <- c(66, 23, 2.6, 0)
# Create configuration object
config <- create_config(file_prefix = "carnivora",
                        data_file = "data/carnivora_deepdive_input.csv",
                        bins = bins,
                        n_areas = length(unique(carnivora$Area)))
# Edit configuration object
config <- edit_config(config = config,
                      module = "general",
                      parameter = "present_diversity",
                      value = 313)
# Save configuration object to file
config$write("carnivora_config.ini")

}
