% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_config.R
\name{edit_config}
\alias{edit_config}
\title{Edit values in the configuration file}
\usage{
edit_config(config = NULL, module = NULL, parameter = NULL, value = NULL)
}
\arguments{
\item{config}{\code{character}. The name of the configuration object, created
using \code{create_config()}, that will be edited.}

\item{module}{\code{character}. The name of the block of the configuration
object to which the parameter belongs.}

\item{parameter}{\code{character}. The name of the parameter to edit in
the configuration file.}

\item{value}{\code{}. The value you want to assign to the parameter. The type
of value is dependent upon the parameter.}
}
\value{
A configuration object with the value of the desired parameter
changed. Once finalised, configuration files should be saved using
\code{config$write("file_name.ini")}.
}
\description{
A function to modify parameters within a configuration object for DeepDive,
created using \code{create_config()}. Default values are listed below, while
the values currently held by each parameter can be checked using the
notation \verb{config$data$[module]$[parameter]}. Extensive checks ensure that
it is possible for the specific parameter to hold the desired value.
}
\details{
A full list of parameters is outlined below.
\emph{general}
\emph{wd} \code{character}. The working directory with which Python should
interact. Defaults to the same as the R working directory, retrieved using
\code{getwd()}.
\emph{time_bins} \code{numeric}. A \code{vector} of bin boundary ages.
\emph{n_areas} \code{integer}. The number of distinct geographic areas across
which occurrences have been sampled. Defaults to 1.
\emph{autotune} \code{logical}. When TRUE (default), a new configuration file
will be saved with adjusted parameters used for analyses that reflect the
empirical data.
\emph{present_diversity} \code{numeric}. The number of species living (extant) in
the modern day. If provided, the simulations are conditioned using this
information. Defaults to NA (meaning this conditioning is not used).

\emph{simulations}
\emph{sim_name} \code{character}. The character string used in the name of the
simulations file. Default is "simulations".
\emph{sims_folder} \code{character}. The name of the directory in which to put the
simulation files. Default is "file_prefix_simulations".
\emph{n_CPUS} <- \code{integer}. The number of CPUs across which to run the
simulations. Default is 1.
\emph{n_training_simulations} \code{integer}. The number of training simulations
to run. Default is 10000.
\emph{training_seed} \code{integer}. The seed number with which to run the
simulations. Default is 123.
\emph{test_seed} \code{integer}. The seed number with which to run the tests. Only
used if \code{add_test = TRUE}. Default is 432.
\emph{n_test_simulations} \code{integer}. The number of test simulations to run.
Only used if \code{add_test = TRUE}. Default is 100.

\emph{s_species} \code{integer}. The number of species with which to start the
simulations. Default is 1.
\emph{total_sp} \code{integer}. The maximum and minimum number of species
permitted in the data set. Default is c(100, 5000).
\emph{root_r} \code{numeric}. The range of ages within which the origin of the
clade is permitted in simulations. Default is
c(0.8 * (max(bins) - min(bins)) + min(bins), max(bins)).
\emph{min_extinct_sp} \code{integer}. The minimum number of extinct lineages
permitted in simulations. Default is 0.
\emph{extant_sp} \code{integer}. The minimum and maximum number of permitted
extant lineages. Default is c(0, 10000).
\emph{rangeL} \code{numeric}. The range of permitted birth rates in simulations.
Default is c(0.02, 0.5).
\emph{rangeM} \code{numeric}. The range of permitted death rates in simulations.
Default is c(0.02, 0.5).
\emph{log_uniform_rates} \code{logical}. If TRUE, birth and death rates are
sampled from a log uniform distribution. Defaults to FALSE.
\emph{p_mass_extinction} \code{numeric}. The probability of a mass extinction
event occurring per million years. Default is 0.01.
\emph{fixed_mass_extinction} \code{numeric}. The age, in millions of years, of a
known mass extinction event to incorporate into simulations. Default is NA
(no known mass extinction).
\emph{magnitude_mass_ext} \code{numeric}. The range of permitted extinction
probabilities during a mass extinction event. Default is c(0.5, 1).
\emph{p_equilibrium} \code{numeric}. The probability of equilibrium per million
years. Default is 0.01.
\emph{p_constant_bd} \code{numeric}. The probability of constant birth and death
rates per million years. Default is 0.01.
\emph{p_mass_speciation} \code{numeric}. The probability of a mass speciation
event per million years. Default is 0.01.
\emph{p_dd_model} \code{numeric}. The probability of diversity-dependent
diversification. Default is 0.05.
\emph{dd_K} \code{numeric}. The carrying capacity for clades under
diversity-dependent diversification. Default is c(100, 1000).
\emph{dd_maxL} \code{numeric}. The starting speciation rate under
diversity-dependent diversification. Default is 1.
\emph{pr_extant_clade} \code{numeric}. The probability of simulating an extant
clade. Default is 0.7.
\emph{poiL} \code{integer}. The expected number of shifts in the birth rate.
Default is 4.
\emph{poiM} \code{integer}. The expected number of shifts in the death rate.
Default is 4.
\emph{scale} \code{integer}. The scaling parameter. Default is 10.
\emph{vectorize} \code{logical}. Should the birth-death simulation be vectorized?
Defaults to TRUE.

\emph{eta} \code{numeric}. Stochasticity in the species-area relationship. Default
is c(1, 1.75).
\emph{p_gap} \code{numeric}. The probability of having no preservation in a
specific time bin. Default is c(0.01, 0.95).
\emph{dispersal_rate} \code{numeric}. The dispersal rate. Default is "None".
\emph{max_dist} \code{integer}. The maximum dispersal distance. Default is 1.
\emph{disp_rate_mean} \code{numeric}. The mean of the distribution defining
dispersal rates. Default is c(0, 1).
\emph{disp_rate_variance} \code{numeric}. The variance of the distribution
defining dispersal rates. Default is 1.
\emph{area_mean} \code{numeric}. The mean of the distribution defining
preservation rates in each geographic area. Default is 20.
\emph{area_variance} \code{numeric}. The variance of the distribution defining
preservation rates in each geographic area. Default is 5.
\emph{size_concentration_parameter} \code{numeric}. The size concentration
parameter. Default is c(0.1, 3).
\emph{link_area_size_carrying_capacity} \code{numeric}. The strength of the
relationship between area size and carrying capacity. Default is c(1, 10).
\emph{p_origination_a_slope_mean} \code{numeric}. The mean slope of the
probability of the mean origination per geographic area. Default is 2.
\emph{p_origination_a_slope_sd} \code{numeric}. The standard deviation of the
slope of the probability of the mean origination per geographic area.
Default is 2.
\emph{sp_mean} \code{numeric}. The mean preservation rates allocated to individual
species. Default is c(0.1, 0.5).
\emph{sp_variance} \code{numeric}. The variance in preservation rates allocated to
individual species. Default is 2.
\emph{slope} \code{numeric}. The change in the log-linear sampling rate through
time. Default is c(-0.01, 0).
\emph{intercept} \code{numeric}. The sampling rate at clade origination. Default
is c(0.1, 0.5).
\emph{sd_through_time} \code{numeric}. The standard deviation in log sampling
rates through time. Default is c(0.001, 0.01).
\emph{sd_through_time_skyline} \code{integer}. The expected number of shifts in
in the sampling rate. Default is 1.
\emph{mean_n_epochs_skyline} \code{integer}. The mean number of epochs. Default is
4.
\emph{fraction_skyline_sampling} \code{numeric}. The fraction of skyline sampling.
Default is 0.5.
\emph{mean_skyline_sampling} \code{numeric}. The mean skyline sampling. Default
is c(0.1, 10).
\emph{maximum_localities_per_bin} \code{integer}. The maximum number of localities
permitted in each time bin. Default is 200.
\emph{species_per_locality_multiplier} \code{numeric}. The multiplier on the
number of species per locality. Default is 1.
\emph{singletons_frequency} \code{numeric}. The proportion of singletons. Default
is 0.1.

\emph{modeltraining}
\emph{sims_folder} \code{character}. The name of the directory containing the
simulation files. Default is "file_prefix_simulations".
\emph{model_folder} \code{character}. The name of the directory in which to put
the model files. Default is "file_prefix_models".
\emph{lstm_layers} \code{integer}. The number of LSTM layers to use. Default is
c(64, 32).
\emph{dense_layer} \code{integer}. The dense layer. Default is c(64, 32).
\emph{dropout} \code{integer}. The dropout. Default is 0.
\emph{max_epochs} \code{integer}. The maximum number of epochs to run. Default is
1000.
\emph{patience} \code{integer}. The patience. Default is 10.
\emph{batch_size} \code{integer}. The batch size. Default is 100.
\emph{validation_split} \code{numeric}. The proportion of data to use for
validation. Default is 0.2.
\emph{f} \code{character}. The name of the feature. Default is
"file_prefix_feature".
\emph{l} \code{character}. The name of the label. Default is
"file_prefix_label".

\emph{empiricalpredictions}
\emph{empirical_input_file} \code{character}. The name of the \code{.csv} file
containing the DeepDive input data, as created using \code{prep_dd_input()}.
\emph{model_folder} \code{character}. The name of the directory containing the
model files. Default is "file_prefix_models".
\emph{n_predictions} \code{integer}. The number of predictions per input file.
Default is 1.
\emph{replicates} \code{integer}. The number of age randomisation replicates
used.
\emph{output_file} \code{character}. The name of the \code{.csv} file to create
containing the DeepDive output. Default is "file_prefix_output".

Parameters relating to geographic area inclusion in the simulations are
better altered using \code{areas_matrix()}.
}
\examples{
# Import internal dataset
data(carnivora)
# Generate vector describing time bin boundaries
bins <- c(66, 23, 2.6, 0)
# Create configuration object
config <- create_config(name = "carnivora",
                        data_file = "data/carnivora_deepdive_input.csv",
                        bins = bins,
                        n_areas = length(unique(carnivora$Area)))
# Edit configuration object
edit_config(config = config, module = "general",
            parameter = "present_diversity", value = 313)

}
