% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localities.R
\name{generate_locality_dataset}
\alias{generate_locality_dataset}
\title{Locality data set}
\usage{
generate_locality_dataset(dat, bins)
}
\arguments{
\item{dat}{Data frame of occurrences with taxon ID, Area, MinAge, MaxAge and
Locality columns.}

\item{bins}{Time bins used in the analysis (should reflect the empirical data
and be identical to the bins used in any corresponding DeepDive simulations
and training)}
}
\value{
A table of dimensions regions x bins where the summed number of
localities in a time bin are recorded.
}
\description{
'generate_locality_dataset()' makes a data frame of the summed number of
localities per area and time bin (output is a table of regions x bins).
}
\examples{
generate_locality_dataset(dat, bins)
}
