% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occs_per_region_time.R
\name{generate_occurrence_dataset}
\alias{generate_occurrence_dataset}
\title{Occurrences per continent and time bin}
\usage{
generate_occurrence_dataset(dat, area_tables, bins)
}
\arguments{
\item{dat}{Data frame of occurrences with taxon ID, Area, MinAge, MaxAge and
Locality columns.}

\item{area_tables}{Tibble of occurrence table split by area}

\item{bins}{Time bins used in the analysis (should reflect the empirical data
and be identical to the bins used in any corresponding DeepDive simulations
and training)}
}
\value{
A table of dimensions regions x bins where the number of occurrences
are summed per region per time bin and recorded.
}
\description{
'generate_occurrence_dataset()' sums the number of occurrences per region to
make a table of regions x bins.
}
\examples{
generate_occurrence_dataset(dat, area_tables, bins)
}
