% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepdive_input_files.R
\name{prep_dd_input}
\alias{prep_dd_input}
\title{Prep DeepDive Input}
\usage{
prep_dd_input(
  dat = dat,
  bins,
  r = replicate,
  age_m = age_method,
  taxon_level = taxonomic_level,
  output_file = NULL
)
}
\arguments{
\item{dat}{Data frame of occurrences with taxon ID, Area, MinAge, MaxAge and
Locality columns.}

\item{r}{The number of age assignment replicates}

\item{age_m}{The age assignment method}

\item{taxon_level}{The taxonomic level the data are identified to}

\item{bin_type}{Determines how bins will be made (equal bins, epochs, stages)}

\item{begin}{Starting age of the oldest time bin}

\item{finish}{Ending age of the youngest time bin}

\item{n_bins}{Number of bins (used only when equal_bins are specified)}

\item{remove_bins}{Specify named time bins to be removed (quote directly or
use a list object)}

\item{use_q}{Merge time bins in the Quaternary together}

\item{merge_holo}{Merge time bins in the Holocene together}

\item{merge_holo_ple}{Merge time bins in the Upper Pleistocene and Holocene}

\item{age_range_threshold}{An age range cut-off between high and low res
occurrences (default=NA)}
}
\value{
Here each age replicate is indicated in the 'R' column and can be
set using the r setting. Locality and occurrence data by time bin are
generated alongside the replicate number, the data type (either locs or occs)
and the region ID.
}
\description{
'prep_dd_input()' makes a single input file for use in DeepDive in one line.
}
\examples{
prep_dd_input(dat=dat, scale="stages", begin=, finish=, r=replicates, age_m = age_method,
taxon_level=taxonomic_level, all_lr=T)
}
