% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepdive_input_files.R
\name{prep_dd_input}
\alias{prep_dd_input}
\title{Prepare DeepDive data input}
\usage{
prep_dd_input(
  dat = NULL,
  bins = NULL,
  r = 1,
  age_m = "median",
  output_file = NULL
)
}
\arguments{
\item{dat}{\code{dataframe}. The \code{dataframe} describing the occurrence data,
including \code{Taxon}, \code{Area}, \code{MinAge}, \code{MaxAge} and \code{Locality} columns. The
supplied \code{dataframe} should not contain any \code{NA} values.}

\item{bins}{\code{numeric}. A numerical \code{vector} designating the boundaries
of the time bins used in the analysis.}

\item{r}{\code{integer}. The number of age assignment replicates. Defaults
to 1.}

\item{age_m}{\code{character}. The age assignment method. Default is
\code{median}. See \code{ages()} for more information.}

\item{output_file}{\code{character}. Defaults to \code{FALSE}, in which case the
DeepDive input table is saved as an object. When a character string is
provided, the object is written to a \code{csv} file with this name.}
}
\value{
If \code{output_file = NULL} a \code{dataframe} containing the occurrence data
summary is returned as an object. If \code{output_file} is given a character
string, the summary is saved to a \code{csv} file with this name, ready to
input into DeepDive.
}
\description{
A function to use fossil occurrence data to make input files for use in
DeepDive.
}
\examples{
# Import internal dataset
data(carnivora)
# Generate vector describing time bin boundaries
bins <- c(66, 23, 2.6, 0)
# Create DeepDive input file
example1 <- prep_dd_input(dat = carnivora, bins = bins)
# Create DeepDive input files using random age assignment, with five reps
example2 <- prep_dd_input(dat = carnivora, bins = bins, r = 5, age_m =
  "random")
}
