% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_occurrences.R
\name{taxa_time_per_area}
\alias{taxa_time_per_area}
\title{Taxa per time bin per area (occurrence data)}
\usage{
taxa_time_per_area(dat, area_tables, bins, taxon_level)
}
\arguments{
\item{dat}{An occurrence data table including taxon id, Area, MinAge, MaxAge
and Locality columns.}

\item{area_tables}{Tibble of occurrence table split by area}

\item{bins}{Time bins used in the analysis (should reflect the empirical data
and be identical to the bins used in any corresponding DeepDive simulations
and training)}

\item{taxon_level}{Specify the taxon level where there is more than one
option}
}
\value{
A table of shape TAXON X REGION X TIME BIN where occurrences are
recorded.
}
\description{
'taxa_time_per_area()' makes a table of shape TAXON X REGION X TIME BIN where
occurrences are recorded.
}
\examples{
taxa_time_per_area(dat, area_tables, bins, taxon_level="Genus")
}
